+++ internal/config/app.go
@@ -29,12 +29,14 @@ type AppConfig struct {
 }
 
 type AppLimit struct {
-	PocketList           int64
-	PocketListItems      int64
-	EventPolaroidLimit   int64
-	LengthPocketListName int64
-	LengthEventName      int64
-	MinimumTopRanking    int64
+	PocketList                     int64
+	PocketListItems                int64
+	EventPolaroidLimit             int64
+	LengthPocketListName           int64
+	LengthEventName                int64
+	LengthEventMessageBoardMessage int64
+	LengthEventAccountingMessage   int64
+	MinimumTopRanking              int64
 }
 
 var APP AppConfig
@@ -70,6 +72,8 @@ func InitialiseConfig() {
 	APP_LIMIT.PocketListItems = 0
 	APP_LIMIT.LengthPocketListName = 0
 	APP_LIMIT.LengthEventName = 0
+	APP_LIMIT.LengthEventMessageBoardMessage = 0
+	APP_LIMIT.LengthEventAccountingMessage = 0
 	APP_LIMIT.MinimumTopRanking = 0
 
 	polaroidLimit, err := strconv.ParseInt(os.Getenv("EVENT_POLAROID_LIMIT"), 10, 64)
@@ -77,6 +81,8 @@ func InitialiseConfig() {
 	pocketListitemsLimit, pocketlistitemsLimitErr := strconv.ParseInt(os.Getenv("POCKET_LIST_ITEMS_LIMIT"), 10, 64)
 	lengthPocketListName, lengthPocketListNameErr := strconv.ParseInt(os.Getenv("LENGTH_POCKET_LIST_NAME"), 10, 64)
 	lengthEventName, lengthEventNameErr := strconv.ParseInt(os.Getenv("LENGTH_EVENT_NAME"), 10, 64)
+	lengthEventMessageBoardMessage, lengthEventMessageBoardMessageErr := strconv.ParseInt(os.Getenv("LENGTH_EVENT_MESSAGE_BOARD_MESSAGE"), 10, 64)
+	lengthEventAccountingMessage, lengthEventAccountingMessageErr := strconv.ParseInt(os.Getenv("LENGTH_EVENT_ACCOUNTING_MESSAGE"), 10, 64)
 	minimumTopRanking, minimumTopRankingErr := strconv.ParseInt(os.Getenv("MINIMUM_TOP_RANKING"), 10, 64)
 	if err == nil {
 		APP_LIMIT.EventPolaroidLimit = polaroidLimit
@@ -95,5 +101,11 @@ func InitialiseConfig() {
 	}
 	if minimumTopRankingErr == nil {
 		APP_LIMIT.MinimumTopRanking = minimumTopRanking
+	}
+	if lengthEventMessageBoardMessageErr == nil {
+		APP_LIMIT.LengthEventMessageBoardMessage = lengthEventMessageBoardMessage
+	}
+	if lengthEventAccountingMessageErr == nil {
+		APP_LIMIT.LengthEventAccountingMessage = lengthEventAccountingMessage
 	}
 }